Reorganize and make data lighter
