+++ snake/tests/cuibm/data/forces
@@ -1,0 +1,10 @@
+0.0004	1412.21	757.725	
+0.0008	2.77228	0.880406	
+0.0012	2.88132	0.867913	
+0.0016	2.79969	0.815889	
+0.002	2.72519	0.772496	
+0.0024	2.64262	0.734236	
+0.0028	2.55261	0.700423	
+0.0032	2.45792	0.671019	
+0.0036	2.36052	0.644569	
+0.004	2.26438	0.622894
