+++ snake/tests/ibamr/data/dataIB/ib_Drag_force_struct_no_0
@@ -1,0 +1,10 @@
+0.00117187	0	0	0	-176.706	5.11374e-08	0
+0.00194282	0	0	0	-124.488	-50.177	0
+0.00253522	0	0	0	-98.0483	-69.0513	0
+0.00302615	0	0	0	-83.6865	-73.0994	0
+0.00345369	0	0	0	-73.3682	-70.5489	0
+0.00383858	0	0	0	-64.5965	-65.237	0
+0.00419339	0	0	0	-56.6918	-58.9125	0
+0.00452638	0	0	0	-49.499	-52.4113	0
+0.00484314	0	0	0	-43.0165	-46.1679	0
+0.00514771	0	0	0	-37.2238	-40.374	0
