+++ snake/tests/openfoam/data/postProcessing/forceCoeffs/0/forceCoeffs.dat
@@ -1,0 +1,19 @@
+# Force coefficients
+# liftDir    : (0.000000e+00 1.000000e+00 0.000000e+00)
+# dragDir    : (1.000000e+00 0.000000e+00 0.000000e+00)
+# pitchAxis  : (0.000000e+00 0.000000e+00 1.000000e+00)
+# magUInf    : 1.000000e+00
+# lRef       : 1.000000e+00
+# Aref       : 1.000000e+00
+# CofR       : (0.000000e+00 0.000000e+00 0.000000e+00)
+# Time       	Cm           	Cd           	Cl           	Cl(f)        	Cl(r)        	Cm	Cd	Cl	Cl(f)	Cl(r)
+2.000000e-03	2.775665e+01	3.511985e+02	3.031645e+02	1.793389e+02	1.238256e+02
+4.000000e-03	-1.732655e-01	3.406270e+00	1.452891e+00	5.531802e-01	8.997112e-01
+6.000000e-03	-3.547398e-01	2.407280e+00	1.070164e+00	1.803423e-01	8.898218e-01
+8.000000e-03	-3.428486e-01	2.161839e+00	1.013337e+00	1.638197e-01	8.495170e-01
+1.000000e-02	-3.319662e-01	2.012603e+00	9.829252e-01	1.594965e-01	8.234288e-01
+1.200000e-02	-3.291763e-01	1.875450e+00	9.549884e-01	1.483179e-01	8.066705e-01
+1.400000e-02	-3.251913e-01	1.767189e+00	9.399970e-01	1.448072e-01	7.951898e-01
+1.600000e-02	-3.186469e-01	1.690922e+00	9.369917e-01	1.498489e-01	7.871428e-01
+1.800000e-02	-3.062728e-01	1.652079e+00	9.484237e-01	1.679390e-01	7.804846e-01
+2.000000e-02	-2.931352e-01	1.628954e+00	9.653689e-01	1.895493e-01	7.758196e-01
