+++ snake/tests/openfoam/data/postProcessing/forces/0/forces.dat
@@ -1,0 +1,13 @@
+# Forces     
+# CofR       : (0.000000e+00 0.000000e+00 0.000000e+00)
+# Time       forces(pressure viscous porous) moment(pressure viscous porous)
+2.000000e-03	((1.748562e+02 1.517026e+02 -4.400584e-19) (7.430454e-01 -1.203620e-01 1.412647e-19) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-7.585131e+01 8.742810e+01 1.390272e+01) (6.018099e-02 3.715227e-01 -2.440008e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+4.000000e-03	((1.150340e+00 8.371695e-01 -9.416888e-21) (5.527951e-01 -1.107239e-01 -7.774773e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-4.185848e-01 5.751700e-01 -5.482078e-02) (5.536193e-02 2.763975e-01 -3.181198e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+6.000000e-03	((7.647149e-01 6.353741e-01 -8.101810e-21) (4.389251e-01 -1.002921e-01 -8.471549e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-3.176871e-01 3.823575e-01 -1.419657e-01) (5.014604e-02 2.194626e-01 -3.540418e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+8.000000e-03	((7.148630e-01 6.022251e-01 -7.985159e-21) (3.660565e-01 -9.555669e-02 -8.158827e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-3.011125e-01 3.574315e-01 -1.341512e-01) (4.777834e-02 1.830283e-01 -3.727314e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+1.000000e-02	((6.897393e-01 5.840187e-01 -7.759839e-21) (3.165620e-01 -9.255611e-02 -7.930562e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-2.920094e-01 3.448697e-01 -1.282577e-01) (4.627806e-02 1.582810e-01 -3.772534e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+1.200000e-02	((6.567477e-01 5.675389e-01 -7.369022e-21) (2.809772e-01 -9.004473e-02 -7.958565e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-2.837695e-01 3.283738e-01 -1.272328e-01) (4.502237e-02 1.404886e-01 -3.735539e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+1.400000e-02	((6.293390e-01 5.580885e-01 -6.965385e-21) (2.542553e-01 -8.808997e-02 -8.065269e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-2.790442e-01 3.146695e-01 -1.259384e-01) (4.404499e-02 1.271276e-01 -3.665722e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+1.600000e-02	((6.118769e-01 5.552053e-01 -6.653254e-21) (2.335839e-01 -8.670951e-02 -8.118081e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-2.776027e-01 3.059385e-01 -1.234784e-01) (4.335475e-02 1.167919e-01 -3.584503e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+1.800000e-02	((6.089494e-01 5.599122e-01 -6.473670e-21) (2.170902e-01 -8.570037e-02 -7.972294e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-2.799561e-01 3.044747e-01 -1.181104e-01) (4.285019e-02 1.085451e-01 -3.502598e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
+2.000000e-02	((6.108744e-01 5.675007e-01 -6.343440e-21) (2.036027e-01 -8.481623e-02 -7.706764e-22) (0.000000e+00 0.000000e+00 0.000000e+00)) ((-2.837503e-01 3.054372e-01 -1.123581e-01) (4.240812e-02 1.018013e-01 -3.420953e-02) (0.000000e+00 0.000000e+00 0.000000e+00))
