+++ snake/tests/petibm/data/forces.txt
@@ -1,0 +1,10 @@
+0.0004	1412.41	758.563
+0.0008	2.71513	0.778018
+0.0012	2.81507	0.801442
+0.0016	2.73036	0.767434
+0.002	2.65867	0.742749
+0.0024	2.57873	0.719169
+0.0028	2.49229	0.696749
+0.0032	2.40291	0.673589
+0.0036	2.31149	0.650686
+0.004	2.22164	0.626377
